update tests about epoch numbers
